Write more functions to accept inputs from the browser.
Most of the choices for the user were taken from the command prompt
(Choice of property set, property card, user, etc.). Now they can be
easily chosen from the browser.

Still some of the inputs are left to be transferred.
TODO: 1. Input for property set, where to insert a property wild card, when
no card is present in the set.
2. Input for whether to play a JSN or DR.
3. Choose button doesn't disappear after the request is sent.
4. Redundant choose buttons appear. Choose button for rent card when no
property card of that color is present.