Refresh userCollection in each giveMoney iter
Now, everytime a player pays a card for ART /ADC / AIB his collection stack
will be refresh so as to remove the card just played. This would prevent
confusion and double clicking on the same card.

Minor bug fix:
Condition for presence of full set for attaching hotels/house updated.